Fix possible need for memory parameter for jobs
Setting default values for memory allocated to jobs prevents the
situation that some jobs cannot start if user does not explicitly ask
for an amount of memory.
This problem was assumed to be caused by the fact that omitting an
explicit request for memory seemed to imply that the job needed all
memory of the node. If any other user had a job running on the node,
this would cause the job to have to wait for any already running jobs
before it could start.